feat(connect): let base-node-rpc detect DropBots
Use new `base_node_rpc` serial device connection code to connect to
DropBot.

If multiple DropBots are detected, automatically connect to DropBot with
highest version with ties going to the first serial port (i.e., `COM1`
before `COM2`).

Offer to *flash firmware* if no DropBot is detected, or if a DropBot is
found with a firmware version that does not match the installed driver
version.