feat(feedback): use thread to watch for threshold
Prior to this commit, the threshold capacitance was periodically checked
using GTK timeout callbacks, i.e., `gobject.timeout_add`.

Instead, notify DropBot of target threshold capacitance and wait for the
DropBot to stream an event indicating the threshold has been met.  Use a
thread and corresponding `threading.Event` instances to watch for the
threshold capacitance (if set) to be reached.  If the threshold is not met
by the specified duration of the step, time out and stop the protocol.