fix(actuated-channels): support any channel
Prior to this commit, an error would occur if a `channels-updated` event
message referenced one or more actuated channels that were not linked not
referenced in the loaded device file (e.g., when running a short test or
channel scan test).

In this commit, fix the area lookup to support channels not in device
(assume an actuated area of 0 for such channels).