fix(events): only emit capacitance if HV enabled
Prior to this commit, the DropBot emitted capacitance update events
regardless of the state of the high-voltage output.  However, the
capacitance measurements are only reliable when the high-voltage output is
enabled.

In this commit, bump the required `dropbot` Conda package version to a
version where capacitance update events are only emitted if high-voltage
output is both **selected** and **enabled**.