fix(threshold): fix target capacitance threshold
Prior to version 2.24, the `max_repeats` step option was used when
`volume_threshold` was specified as follows:

 1. Apply actuation for `duration`
 2. If measured capacitance > target threshold, complete step.
 3. Repeat 1-3 up to `max_repeats` times.

As of version 2.24, the `max_repeats` step option is no longer used.
Instead, use the following procedure:

 1. Apply actuation and submit request to DropBot to notify when target
    capacitance is exceeded.
 2. Wait for `capacitance-exceeded` event is received.
 3. If `duration` passes while waiting, notify that the step has failed.

The >=2.24 method is preferable since the protocol step will be notified
as complete as soon as the target capacitance is reached, whereas in the
<2.24 method, at least `duration` time is taken.  However, in the event
the DropBot does not emit a `capacitance-exceeded` event, the protocol
will fail _even if the target capacitance was was reached_.

In this commit,	connect to the `capacitance-updated` event (in addition to
`capacitance-exceeded` event) to check against target capacitance.