chore(events): do not use capacitance-exceeded
Prior to this commit, the `capacitance-exceeded` DropBot event was
connected to a callback waiting on a threshold capacitance to be met.
However, [an issue with the DropBot `capacitance-exceeded` event][i32]
caused corresponding JSON messages to be truncated.

In this commit, as a workaround until this issue is resolved, do not use
the `capacitance-exceeded` DropBot event.  Instead, only use
`capacitance-updated` messages to detect when a target capacitance has
been reached.

[i32]: https://gitlab.com/sci-bots/dropbot.py/issues/32