refactor(capacitance/log): remove sampling_rate_hz
Prior to this commit, logged capacitance updates contained a column named
`sampling_rate_hz`.  However, this was _not_ actually the analog sampling
rate, but rather a calculated _effective_ sampling rate based on the time
taken to acquire the analog samples and compute the capacitance (which
involves sorting).

In this commit, remove the `sampling_rate_hz` column from the capacitance
log table, since it may have lead to confusion.  Also, move the
`capacitance` column to index 1 (adjacent to the `timestamp_utc` column),
to make, e.g., plotting in Excel easier.