feat(device-state): add `_state_applied` event
The DropBot state is set in the `DropBotPlugin._apply_state` method.
Prior to this commit, a race condition existed where it was possible that
the `_wait_for_target_capacitance` function could be called before the
updated list of actuated channels was reported from the DropBot in a
`channels-updated` event.  Such cases resulted in undefined behaviour
since the list of actuated channels used for the `actuated_area`
calculation was stale (i.e., not up to date).

In this commit, add the `_state_applied` Python threading event.  Before
computing target capacitance based on electrode areas for a step wait for
the `_state_applied` event to ensure that pending electrode actuations
have completed.