feat(firmware-flash): ignore patch number mismatch
Prior to this commit, any mismatch between the DropBot driver and firmware
versions resulted in prompting the user to flash the matching firmware
version.

In this commit, ignore mismatches between DropBot driver and firmware
versions as long as base minor versions are equal, i.e., assume API is
backwards compatible.