feat(channels): add channel states event
Prior to this commit, for each step, actuation was applied before the
electrode controller plugin could notify the DropBot plugin of the
electrodes to actuate.  This resulted in the electrodes from the previous
step being actuated, followed by a second actuation operation by the
DropBot plugin to handle the updated step electrodes sent from the
electrode controller plugin.

In this commit, add `_channel_states_received` event, which is cleared at
the start of each step and is set once electrode states have been received
from the electrode controller plugin.