perf(step-run): volume threshold -> duration=0
Prior to this commit, during protocol execution, a step was not completed
until the specified step duration had elapsed, regardless of whether a
DropBot was connected or not.
When a DropBot **_was_** connected and a volume threshold was specified for
a step, the `duration` step parameter was interpreted as the *maximum*
duration of the step, i.e., the timeout period to wait for the specified
threshold to be reached.  However, when a DropBot was *not* connected, the
volume threshold is never reached, so the duration can be quite long.

In this commit, if a DropBot is not connected during protocol execution
and a volume threshold is specified, treat the `duration` step parameter
as a maximum step duration and continue immediately.  If no volume
threshold is specified, maintain the behaviour from prior to this commit,
i.e., delay for the specified duration.