feat(step-run): execute step in separate thread
Prior to this commit, step handling was executed in a separate thread.
This prevented other GTK events and step handling of other plugins from
executing concurrently.  While this not a problem in general, it may
present an issue, for example, if the DropBot plugin needs to coordinate
with another plugin.

In this commit, execute step handling in a separate thread.