fix(step-cancel): only cancel current step
Prior to this commit, delayed callbacks using the step duration may have
been executed after switching to a different step in the protocol.  This
could lead to the incorrect step being cancelled by a callback.

In this commit, assign each step execution a UUID, and verify the UUID
against the currently running step in each callback.  This ensures that
cancel callback operations are only carried out during the step from which
they were launched.