fix(label): fix chip inserted status
Prior to this commit and as of version 2.26, the chip-inserted status
does not update as expected.  Although the chip inserted status is set
correctly in the status label upon initial connection to the DropBot, the
status is not updated when the chip is inserted/removed afterwards.

In this commit, connect to the `output_enabled` and `output_disabled`
DropBot signals to update the chip insertion status.