perf(label): refresh capacitance every >= 250 ms
Prior to this commit, the capacitance in the status label was updated
every time a `capacitance-udpated` event was received from the DropBot.
However, this may result in poor performance for small capacitance update
intervals, e.g., < 100 ms.

In this commit, only refresh the capacitance in the GUI status label *at
most* every 250 ms.