refactor(signals): move handlers init from connect
Prior to this commit, DropBot signal handlers were connected when the
`dropbot-connected` `gobject` signal was emitted.  However, this prevented
the `dropbot-connected` signal from being reetrant, i.e., could not be
called more than once, since the DropBot signal handlers would be
connected multiple times causing them to unexpectedly be called multiple
times when the respective signals were sent by the DropBot.  Ideally, the
`dropbot-connected` signal should be emitted whenever a connection to the
DropBot is established, _including_ when a connection is _re_established.
This provides an opportunity for the run-time state of the DropBot to be
initialized to a known configuration.

In this commit, refactor the `dropbot-connected` callback to make it
reentrant.  This involves separating connection of the the DropBot signal
handlers into a separate function that is _only_ called upon the _initial_
connection to the DropBot.