fix(capacitance): update logging for new message
Prior to this commit, the logging of capacitance values at the end of a
step assumed that `capacitance-updated` messages included `start` and
`end` properties.  However, as of `dropbot >=1.54`, the `start` property
was dropped from the `capacitance-updated` messages and the `end` property
was renamed to `time_us`.

In this commit, update the logging of capacitance values to use the
renamed `time_us` event property.