refactor(commands): register capacitance measuring
Prior to this commit, requests for measuring liquid or filler capacitance
were handled by subscribing to the corresponding messages from the
`dmf_device_ui_plugin` on the ZeroMQ plugin interface.  However, as of
`dmf-device-ui >=0.13`, the capacitance measurement commands were removed
in favour of dynamically generating a context menu based on the commands
registered with the `microdrop.command_plugin`.

In this commit, do not subscribe to messages on the ZeroMQ interface and
register commands to measure liquid capacitance and filler capacitance
with the `microdrop.command_plugin`.