chore(app-fields): disable auto-tests by default
Prior to this commit, diagnostic tests were run automatically by default
whenever the experiment log was changed (e.g., when a new experiment was
started).  However, there was no indication provided to the user as to
what was happening and the UI appeared to freeze.  This was particularly
noticeable when shutting down MicroDrop, where the UI would freeze while
the auto-tests were run before closing the app.  It should also be noted
that, as of this commit, short detection is automatically run a) whenever
a chip is inserted into the DropBot; and b) whenever serial connection to
the DropBot is (re)established; making automatic execution of diagnostic
tests somewhat less important.

In this commit, disable the automatic diagnostic tests by default.

See #4 for proposed change to run auto-diagnostic tests in cancellable
coroutine and display progress, which would make execution of the tests
less intrusive.