fix(protocol-run): do not warn about no connection
Prior to this commit, if a protocol was started and a DropBot was not
connected, a warning dialog was displayed to notify the user and the
protocol continued anyway once the user clicked OK.

In this commit, do not warn about not DropBot connection when a protocol
is run.  As of MicroDrop 2.28.2, a warning is displayed by the electrode
controller plugin in such cases with an option to ignore the issue and
continue.