feat(step-run): reduce wait w/no DropBot connected
Prior to this commit, if no DropBot was connected while executing a step,
each actuation would delay for the entire specified duration, regardless
of the volume threshold setting.  However, when the DropBot _is_
connected, the expected behaviour is that specified duration is the
_maximum_ time to wait for volume threshold to be reached.

In this commit, rather than wait the full specified duration for each
actuation, temporarily set the duration to 0.2 seconds during the step to
mimic typical maximum actuation speed, e.g., when volume threshold is set
with the test board inserted.