fix(status): ignore infinite liquid C
Prior to this commit, conducting liquid capacitance calibration on an
empty electrode could result in an _infinite_ calculated specific
capacitance.  This caused an exception when formatting the specific
capacitance for the main window status label.

In this commit, ignore specific capacitance values that are infinite when
formatting the main window status label.