refactor(protocol): ref ctrl for protocol state
Prior to this commit, the state of the running protocol (e.g., current
step number) was stored directly in the `app.protocol` object.  However,
as of Microdrop v2.33, the state of the running protocol is managed by the
`app.protocol_controller` object instead.

In this commit, update references to the running protocol state to use the
`app.protocol_controller` object.