fix(halted): fix halted error dialog
Prior to this commit, a typo caused an error when a "halted" event was
reported by the DropBot, e.g.:

```python
TypeError: sequence item 1: expected string, list found
```

In this commit, fix they typo to show an error dialog to the user
indicating the cause of the "halted" event.

Fixes #5.