feat(halted): detect shorts on run/real-time mode
Prior to this commit, channels disabled due to a "halted" event reported
by the DropBot remained disabled until the user reset the DropBot control
board microcontroller, e.g., by unplugging all cables and plugging them
back in.

In this commit, When real-time mode is enabled or when a protocol starts
running, detect shorts to enable any channels where a short is not
detected.  This, e.g., automatically re-enables channels after a "halted"
event.