docs(pyutilib): note about experiment controller
Prior to this commit, as of version 2.38.5, the
`DropBotPlugin.data_dir()` method used the `experiment_ctrl` attribute
of the `ExperimentLogController` plugin to query the data directory for
the current experiment.  However, this changes introduced in
`microdrop==2.34`, which are still considered experimental and have not
been accepted for public release.

In this commit, add a note to indicate when this change was made.