feat(pyutilib): debug log capacitance filename
Prior to this commit, the path of the capacitance updates log file was
only logged after successfully writing to the log file. However, in
cases where logging capacitance updates fails to complete successfully,
it may be useful to know the intended path of the log file.

In this commit, also log the _planned_ path of the capacitance log file
before attempting to log capacitance updates.